Add v1beta1 versions of Constraints and ConstraintTemplates
Note that this also required moving to a model where we use a common
core representation of the template resource and implement a method for
conversion between each version.

Another issue was that the version of controller-gen that ships
with Kubebuilder v1 doesn't handle multi-version CRDs, which I've
addressed by patching the versions in via kustomize.

Signed-off-by: Max Smythe <smythe@google.com>