Preserve unknown fields in ByPodStatus
A Gatekeeper e2e test is currently failing in
open-policy-agent/gatekeeper#1286 due to the use of the `operations`
field.  This field is added by Gatekeeper, under a ConstraintTemplate's
`byPodsStatus`.  This field relied on the v1beta1 CRD behavior, where
unknown fields are preserved by default.

As this field fundamentally has to do with gatekeeper, it's better not
to define it in the Constraint Framework.  Other CF consumers have no
need for the Operations field.

Even better, we may eventually define ConstraintTemplateSpec in
Constraint Framework consumers, rather than in the library itself.

For now, adding `x-kubernetes-preserve-unknown-fields: true` to
ByPodStatus will solve our problem and maintain the existing behavior.

Contributes to open-policy-agent/gatekeeper#550

Signed-off-by: juliankatz <juliankatz@google.com>