Inject empty `parameters` field for legacy schemas
To handle legacy (pre-v1) CRDs, we inject
x-kubernetes-preserve-unknown-fields
in order to preserve the legacy behavior
that unknown fields were not pruned by the API
server.

Unfortunately, we are not handling the case
where no parameters are provided, which currently
means that parameters will be stripped from users' CRs.

This fixes that.

Fixes: https://github.com/open-policy-agent/gatekeeper/issues/1468

Signed-off-by: Max Smythe <smythe@google.com>