Provide IsUnrecognizedConstraintError
For decoupling frameworks and gatekeeper error handling
(which allows using go113 error conventinos in tests), I need to remove
the dependence of gatekeeper on error types defined in frameworks so
these may be changed safely to raw errors (where reasonable to do so).

As an intermediate, I'm adding this function so that gatekeeper can
continue to compile, and the implementation of this method may be
switched out in frameworks.

Signed-off-by: Will Beason <willbeason@google.com>