Make Client E2E tests more flexible (#150)
* Rename e2e_tests to e2e_tests

Signed-off-by: Will Beason <willbeason@google.com>

* Make tests more flexible

Refactor tests so that:

- it is possible to specify rego and libraries per-test
- all tests use their own Client
- the Probe type is no longer necessary

Signed-off-by: Will Beason <willbeason@google.com>

* Use cmp.Diff to check for response differences

This makes test output on failure more consistent and clearer.

Use per-test Context instead of one shared between all tests. This
removes another avenue of test crosstalk and allows us to test things
like behavior with cancellable contexts.

Make response order more, but not completely, deterministic. We don't
need response order to be more deterministic for tests than what I've
put here, and as-is will make the output of reviewing objects more
consistent. There are common edge cases where ordering is randomized,
but this can be fully fixed later.

Comment out the contents of TestRemoteClientE2E. This test is currently
skipped until we have a working remote to connect the remote Driver to.
I've added clarifying comments to explain what we'll want once this
exists.

Signed-off-by: Will Beason <willbeason@google.com>

* Make reviewer-suggested changes

Signed-off-by: Will Beason <willbeason@google.com>