Benchmark AddConstraint function
We didn't expect this to be problematic, but it's nice to have a
benchmark for this as we do intend to make some changes to what happens
when Constraints are added.

Fix "ModuleSimple" so it actually compiles. I was originally against
having the error check in the benchmark since it's slightly slower, but
(1) the performance difference is negligible and (2) we want to be sure
the CT is actually being compiled, so it's worth the minuscule
performance cost.

Signed-off-by: Will Beason <willbeason@google.com>