refactor: Remove client.Reset (#169)
* refactor: Remove client.Reset

Now that gatekeeper no longer uses it, there are no callers. All tests
have previously been migrated off to avoid potential crosstalk. It's
best to wait until we have solid plans for a remote implementation to
consider modifying the design of Client.

This makes the implementation work for reo compilation
sharding/optimization significantly easier.

Signed-off-by: Will Beason <willbeason@google.com>

* Remove DeleteModule

This is no longer used so it can be safely deleted.

Signed-off-by: Will Beason <willbeason@google.com>