Remove direct storage.Path requirement in handler (#208)
* Remove direct storage.Path requirement in handler

As-is requires a direct dependency of all callers on OPA. This PR
changes the TargetHandler interface to accept storage path as a slice of
strings instead of storage.Path.

Functionally this changes nothing.

Signed-off-by: Will Beason <willbeason@google.com>

* Remove storage.Path from Cacher

Signed-off-by: Will Beason <willbeason@google.com>

* Upgrade golangci-lint

The workflow seems broken? Upgrading to fix.

Signed-off-by: Will Beason <willbeason@google.com>

* Update make lint rule

Signed-off-by: Will Beason <willbeason@google.com>

* Incrase timeout for gatekeeper tests.

We've been at the threshold of about 5 minutes for a while, causing
flaky tests. We recently enabled --race, which slows all tests and is
likely what makes the tests run too slow in GitHub actions to run within
the current 5 minute timeout.

Note that in order to run, this target must create an entirely new
Gatekeeper docker image with every invocation - Docker does not cache
steps which involve file operations, so the image must be built from the
ground up each time. This alone takes up a couple minutes of the
runtime.

As stated in the PR discussion, I see no performance degradation in the
frameworks tests. I've had trouble getting Gatekeeper tests to pass at
all from frameworks in GitHub, but every run I've done locally has
worked.

Signed-off-by: Will Beason <willbeason@google.com>